fix: correct placement for workflow context
