+++ packages/app-headless-cms-workflows/src/index.tsx
@@ -1,13 +1,17 @@
 import React from "react";
+import { Wcp } from "@webiny/app-admin";
 import { ContentEntryFormWorkflow } from "~/Components/ContentEntryFormWorkflow.js";
 import { CmsWorkflowsEditor } from "~/Views/CmsWorkflowsEditor.js";
-import { Wcp } from "@webiny/app-admin";
+import { CmsEntryFormTooltipButton } from "~/Components/CmsEntryFormTooltipButton.js";
+import { ContentEntryWorkflow } from "~/Components/ContentEntryWorkflow.js";
 
 export const CmsWorkflows = () => {
     return (
         <Wcp.CanUseWorkflows>
             <CmsWorkflowsEditor />
             <ContentEntryFormWorkflow />
+            <ContentEntryWorkflow />
+            <CmsEntryFormTooltipButton />
         </Wcp.CanUseWorkflows>
     );
 };
